undo unneeded changes
